% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABBREV}
\alias{ABBREV}
\title{Contains all the food}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8789 rows and 53 columns.}
\usage{
ABBREV
}
\description{
Contains all the food items found in the relational database, but with fewer nutrients and other related information. Data refer to 100 g of the edible portion of the food item.
}
\keyword{datasets}

