% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_nutrients}
\alias{calculate_nutrients}
\title{Calculates nutrients}
\usage{
calculate_nutrients(ingredients, grams)
}
\arguments{
\item{ingredients}{is a vector of one or more ingredients}

\item{grams}{is a vector of the corresponding weight in grams of each ingredient}
}
\value{
a data frame containing the nutrients of each ingredient along with a total score.
}
\description{
The function calculates nutrient value based on a list of ingredients and their weight in grams.
}
\examples{
library(NutrientData)
library(dplyr)

ingredients <- c("CABBAGE,RAW", "MAYONNAISE,RED FAT,W/ OLIVE OIL", "ONIONS,RAW")
grams <- c(100, 20, 10)

calculate_nutrients(ingredients, grams) \%>\%
  select(Food = 1, Calories = 3, Protein = 4,
         Fat = 5, Carbs = 7)

}

