% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_SRC}
\alias{DATA_SRC}
\title{Sources of Data File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 682 rows and 9 columns.}
\usage{
DATA_SRC
}
\keyword{datasets}

