% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATSRCLN}
\alias{DATSRCLN}
\title{Sources of Data Link File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 244496 rows and 3 columns.}
\usage{
DATSRCLN
}
\description{
This file is used to link the Nutrient Data file with the Sources of Data table. It is needed to resolve the many-to-many relationship between the two tables.#' \itemize{
  \item Nutrient Data file by NDB No. and Nutr_No
  \item Nutrient Definition file by Nutr_No
  \item Sources of Data file by DataSrc_ID
}
}
\keyword{datasets}

