% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DERIV_CD}
\alias{DERIV_CD}
\title{Data Derivation Code Description File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 2 columns.}
\usage{
DERIV_CD
}
\description{
This file provides information on how the nutrient values were determined. The file contains the derivation codes and their descriptions.#' Links to:
\itemize{
  \item Nutrient Data file by Deriv_Cd
}
}
\keyword{datasets}

