% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FD_GROUP}
\alias{FD_GROUP}
\title{Food Group Description File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 2 columns.}
\usage{
FD_GROUP
}
\description{
This file is a support file to the Food Description file and contains a list of food groups used in SR28 and their descriptions.
Links to:
\itemize{
  \item Food Description file by FdGrp_Cd
}
}
\keyword{datasets}

