% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOOD_DES}
\alias{FOOD_DES}
\title{Food Description File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8789 rows and 14 columns.}
\usage{
FOOD_DES
}
\description{
This file contains long and short descriptions and food group designators for all food items, along with common names, manufacturer name, scientific name, percentage and description of refuse, and factors used for calculating protein and kilocalories, if applicable. Items used in the FNDDS are also identified by value of “Y” in the Survey field.
Links to:
\itemize{
  \item Food Group Description file by the FdGrp_Cd field
  \item Nutrient Data file by the NDB_No field
  \item Weight file by the NDB_No field
  \item Footnote file by the NDB_No field
  \item LanguaL Factor file by the NDB_No field
}
}
\keyword{datasets}

