% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOOTNOTE}
\alias{FOOTNOTE}
\title{Footnote File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 553 rows and 5 columns.}
\usage{
FOOTNOTE
}
\description{
This file contains additional information about the food item, household weight, and nutrient value.
\itemize{
  \item Food Description file by NDB_No
  \item Nutrient Data file by NDB_No and when applicable, Nutr_No
  \item Nutrient Definition file by Nutr_No, when applicable
}
}
\keyword{datasets}

