% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LANGDESC}
\alias{LANGDESC}
\title{LanguaL Factors Description File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 774 rows and 2 columns.}
\usage{
LANGDESC
}
\description{
This file is a support file to the LanguaL Factor file and contains the descriptions for only those factors used in coding the selected food items codes in this release of SR.
Links to:
\itemize{
  \item LanguaL Factor file by the Factor_Code field
}
}
\keyword{datasets}

