% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LANGUAL}
\alias{LANGUAL}
\title{LanguaL Factor File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38301 rows and 2 columns.}
\usage{
LANGUAL
}
\description{
This file is a support file to the Food Description file and contains the factors from the LanguaL Thesaurus used to code a particular food.
Links to:
\itemize{
  \item Food Description file by the NDB_No field
  \item LanguaL Factors Description file by the Factor_Code field
}
}
\keyword{datasets}

