% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NUT_DATA}
\alias{NUT_DATA}
\title{Nutrient Data File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 679045 rows and 18 columns.}
\usage{
NUT_DATA
}
\description{
This file contains the nutrient values and information about the values, including expanded statistical information.
#' Links to:
\itemize{
  \item Food Description file by NDB_No
  \item Food Description file by Ref_NDB_No
  \item Weight file by NDB_No
  \item Footnote file by NDB_No and when applicable, Nutr_No
  \item Sources of Data Link file by NDB_No and Nutr_No
  \item Nutrient Definition file by Nutr_No
  \item Source Code file by Src_Cd
  \item Data Derivation Code Description file by Deriv_Cd
}
}
\keyword{datasets}

