% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NUTR_DEF}
\alias{NUTR_DEF}
\title{Nutrient Definition File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 6 columns.}
\usage{
NUTR_DEF
}
\description{
This file is a support file to the Nutrient Data file. It provides the 3-digit nutrient code, unit of measure, INFOODS tagname, and description.#'
Links to:
\itemize{
  \item Nutrient Data file by Nutr_No
}
}
\keyword{datasets}

