% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_ingredient}
\alias{search_ingredient}
\title{Searches the data}
\usage{
search_ingredient(ingredient)
}
\arguments{
\item{ingredient}{is a string of one or more comma-separated terms}
}
\value{
a vector of possible matches
}
\description{
The function takes one or more comma-separated search terms and looks for matches in the ABBREV data set.
}
\examples{
library(NutrientData)

search_ingredient("onion")

search_ingredient("onion,raw")

}

