% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRC_CD}
\alias{SRC_CD}
\title{Source Code File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.}
\usage{
SRC_CD
}
\description{
This file contains codes indicating the type of data (analytical, calculated, assumed zero, and so on) in the Nutrient Data file. To improve the usability of the database and to provide values for the FNDDS, NDL 34staff imputed nutrient values for a number of proximate components, total dietary fiber, total sugar, and vitamin and mineral values.
Links to:
\itemize{
  \item Nutrient Data file by Src_Cd
}
}
\keyword{datasets}

