% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WEIGHT}
\alias{WEIGHT}
\title{Weight File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15438 rows and 7 columns.}
\usage{
WEIGHT
}
\description{
This file  contains the weight in grams of a number of common measures for each food item.
\itemize{
  \item Food Description file by NDB_No
  \item Nutrient Data file by NDB_No
}
}
\keyword{datasets}

